Pull out SignRequest method.
